==> 9_association_with_top_snp.Rmd <==
---
title: "Untitled"
author: "Takeo Naito, MD PhD"
date: "12/1/2020"
output: html_document
---



## export genotyping info for top snps
```{bash}
plink2 \
--bfile /mnt/isilon_data/For_Takeo2/From_nick/Ichip/plinkfile/ichip_pancreatitis_for_PCs \
--export a \
--snp rs2647087 \
--out /mnt/share6/FOR_Takeo/hla/top_snp
```

```{r}
load("/mnt/share6/SHARED_DATASETS/hla_working_group/results/hla_table.RData")

hla_table <- hla_table %>% 
  mutate(FID = as.character(FID))

```

```{r}
pheno = fread("/mnt/share6/For_Takeo/phenotypdata/pancreas_subclinical_phenotypes.txt")
pheno <- pheno %>% 
  mutate(FID = str_replace_all(GeneticID,"-","0"))

pheno <- pheno %>% 
  select(FID,pancreatitis,Diag)

pheno <- pheno %>% 
  mutate(pancreatitis = ifelse(pancreatitis == 1,0,
                               ifelse(pancreatitis == 2,1,NA)))
```

```{r}
pc <- fread("/mnt/share6/FOR_Takeo/ICHIP/PC_ichip1to8washubbc_cauc.txt")
```

```{r}
hla_v2 <- hla_table %>% 
  left_join(pheno, by = "FID")

hla_v3 <- hla_v2 %>% 
  left_join(pc,by = "FID")


hla_v3 <- hla_v3 %>% 
  filter(!is.na(pancreatitis))
```

```{r}
relate = fread("/mnt/isilon_data/For_Takeo2/From_nick/Ichip/related_ichip")


hla_v3 <- hla_v3 %>% 
  filter(!FID %in% relate$FID) 


hla_v3 <- hla_v3 %>% 
  filter(!is.na(PC1))

hla_v3 <- hla_v3 %>% 
  filter(Diag %in% c("CD","UC","IBDU"))
```

```{r}
AJ = fread("/mnt/share6/FOR_Takeo/hla/hla_AJ_list")
AJ = AJ %>% 
  mutate(group = "AJ")

non_AJ = fread("/mnt/share6/FOR_Takeo/hla/hla_nonAJ_list")

non_AJ = non_AJ %>% 
  mutate(group = "non_AJ")

AJ_merge = rbind(AJ,non_AJ)


hla_v4 = hla_v3 %>% 
  select(-starts_with("HLA")) %>%
  left_join(AJ_merge,by = "FID")

hla_nonaj = hla_v3 %>% 
  select(-starts_with("HLA")) %>% 
  filter(FID %in% non_AJ$FID)


```



## read genotypes

```{r}
geno = fread("/mnt/share6/FOR_Takeo/hla/top_snp.raw")
geno = geno %>% 
  mutate(top_snp = ifelse(rs2647087_A == 2, 0,
                          ifelse(rs2647087_A ==0,2,rs2647087_A)))

hla_v5 <- hla_v4 %>% 
  left_join(geno %>% select(FID,top_snp),by = "FID")


```

## do regression 
```{r}
hla_aj = hla_v5 %>% filter(group == "AJ")
hla_nonaj = hla_v5 %>% filter(group == "non_AJ")
fulmodel = glm(pancreatitis ~ top_snp + PC1 + PC2 + PC3 + PC4,data = hla_v5,family = "binomial")

summary(fulmodel)


aj_model = glm(pancreatitis ~ top_snp + PC1 + PC2 + PC3 + PC4,data = hla_aj,
               family = "binomial")
summary(aj_model)

non_aj_model = glm(pancreatitis ~ top_snp + PC1 + PC2 + PC3 + PC4,data = hla_nonaj,
                   family = "binomial")
summary(non_aj_model)

```


```{r}
alt_num_cont = hla_v5 %>% filter(pancreatitis == 0 & top_snp == 2) %>% nrow*2 +
      hla_v5 %>% filter(pancreatitis == 0 & top_snp == 1) %>% nrow() 

num_all_cont= hla_v5 %>% filter(pancreatitis == 0 & !is.na(top_snp)) %>% nrow()*2
    
maf_cont = alt_num_cont/num_all_cont
maf_cont
    
    
alt_num_case = hla_v5 %>% filter(pancreatitis == 1 & top_snp == 2) %>% nrow*2 +
      hla_v5 %>% filter(pancreatitis == 1 & top_snp == 1) %>% nrow() 
num_all_case= hla_v5 %>% filter(pancreatitis == 1 & !is.na(top_snp)) %>% nrow()*2
    
maf_case = alt_num_case/num_all_case
maf_case
    
```


```{r}

alt_num_cont_aj = hla_aj %>% filter(pancreatitis == 0 & top_snp == 2) %>% nrow()*2 +
      hla_aj %>% filter(pancreatitis == 0 & top_snp == 1 ) %>% nrow() 

num_all_cont_aj= hla_aj  %>% filter(pancreatitis == 0 & !is.na(top_snp)) %>% nrow()*2
    
maf_cont_aj = alt_num_cont_aj/num_all_cont_aj
maf_cont_aj
    
    
alt_num_case_aj = hla_aj %>% 
  filter(pancreatitis == 1 & top_snp == 2) %>% nrow()*2 +
      hla_aj %>% filter(pancreatitis == 1 & top_snp == 1) %>% nrow() 

num_all_case_aj= hla_aj %>% filter(pancreatitis == 1 & !is.na(top_snp)) %>% nrow()*2
    
maf_case_aj = alt_num_case_aj/num_all_case_aj
maf_case_aj


  

```


```{r}
alt_num_cont_nonaj = hla_nonaj %>% filter(pancreatitis == 0 & top_snp == 2) %>% nrow()*2 +
      hla_nonaj %>% filter(pancreatitis == 0 & top_snp == 1) %>% nrow() 

num_all_cont_nonaj= hla_nonaj  %>% filter(pancreatitis == 0 & !is.na(top_snp)) %>% nrow()*2
    
maf_cont_nonaj = alt_num_cont_nonaj/num_all_cont_nonaj
maf_cont_nonaj
    
    
alt_num_case_nonaj = hla_nonaj %>% 
  filter(pancreatitis == 1 & top_snp == 2) %>% nrow()*2 +
      hla_nonaj %>% filter(pancreatitis == 1 & top_snp == 1) %>% nrow() 

num_all_case_nonaj= hla_nonaj %>% filter(pancreatitis == 1 & !is.na(top_snp)) %>% nrow()*2
    
maf_case_nonaj = alt_num_case_nonaj/num_all_case_nonaj
maf_case_nonaj
```



```{bash}
plink2 \
--bfile /mnt/isilon_data/For_Takeo2/From_nick/Ichip/plinkfile/ichip_pancreatitis_for_PCs \
--freq \
--snp rs2647087 \
```
